refactor: 데이터 랜덤 생성 함수 일부를 이전
